requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorclass==2.2.2
cron_descriptor==2.0.6
Django==5.2.11
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-crontab==0.7.1
django-timezone-field==7.2.1
djangorestframework==3.16.1
docopt==0.6.2
frozenlist==1.8.0
gunicorn==25.0.3
honcho==2.0.0
idna==3.11
kombu==5.6.2
multidict==6.7.1
packaging
pillow==12.1.0
pip-upgrader==1.4.15
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.1
requests==2.32.5
setuptools==80.10.2
six==1.17.0
sqlparse==0.5.5
terminaltables==3.1.10
twilio==9.10.1
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
wheel==0.46.3
whitenoise==6.11.0
yarl==1.22.0
